requirements.txt
PyPDF2 
python-docx
pytesseract
pdf2image
Pillow 
langdetect 
google-genai
PyMuPDF
python-dotenv
tqdm
fastapi[standard]